converted to sp, no codegen yet

+++ src/structures.h
@@ -2,13 +2,13 @@
 #define STRUCTURES_H
 
 #ifndef __CUDACC__
-  struct double3 { double x, y, z; };
+  struct float3 { float x, y, z; };
 #else
   // nvcc will pull in '/usr/include/vector_types.h' which 
-  // contains an identical definition of 'double3' (above).
+  // contains an identical definition of 'float3' (above).
 #endif
 
-struct edge { double a, b; };
-struct edge_neighbour { int a, b; double x, y, z; };
+struct edge { float a, b; };
+struct edge_neighbour { int a, b; float x, y, z; };
 
 #endif
